my-email.el: slightly better path setting
We need to ensure the path is right for compile time as well and it
varies depending on the machine. Not all mu4e requires are available on
travis VMs which have to make do with the package.

I've made other hacks to deal with older mu4e's in Travis but I suspect
the best solution is to get the updated mu4e packaged.