my-c-mode: fix lsp-mode invocation
I was using the wrong hook function, M-x lsp will import a workspace
and then things start working. I might consider expanding it from
c-mode to prog-mode as per the suggestion of:

  https://raw.githubusercontent.com/emacs-lsp/lsp-mode/master/scripts/lsp-start-plain.el

which is a known working configuration. Also take the time to clean-up
the settings, remove dead variables and enable the faster plist mode.